+++ include/lingodb/compiler/frontend/sql_scope.h
@@ -8,22 +8,14 @@
 #include <memory>
 namespace lingodb::ast {
 class BoundExpression;
-}
+} // namespace lingodb::ast
 namespace lingodb::analyzer {
 
 class SQLScope {
    public:
-   std::map<std::string, std::pair<size_t, ast::NamedResult>> tables;
-   std::map<std::string, std::vector<std::pair<size_t, ast::NamedResult>>> columns;
-   std::map<std::string, std::vector<std::pair<size_t, ast::NamedResult>>> functionsEntry;
    //Stores the columns that should be produced at the end
    ast::TargetInfo targetInfo;
-   std::vector<std::shared_ptr<ast::BoundExpression>> evalBefore;
-
-   //Stores the grouping functions
-   std::vector<std::shared_ptr<ast::BoundFunctionExpression>> groupingFunctions;
 
    std::shared_ptr<SQLScope> parent;
-   size_t id = 0;
 };
 } // namespace lingodb::analyzer
