+++ src/compiler/frontend/ast/bound/bound_expression.cpp
@@ -3,7 +3,7 @@ namespace lingodb::ast {
 /*
  * BoundColumnRefExpression
 */
-BoundColumnRefExpression::BoundColumnRefExpression(std::string scope, catalog::NullableType resultType, std::shared_ptr<NamedResult> namedResult, std::string alias) : BoundExpression(TYPE, ExpressionType::BOUND_COLUMN_REF, resultType, alias), scope(scope) {
+BoundColumnRefExpression::BoundColumnRefExpression(catalog::NullableType resultType, std::shared_ptr<NamedResult> namedResult, std::string alias) : BoundExpression(TYPE, ExpressionType::BOUND_COLUMN_REF, resultType, alias) {
    this->namedResult = namedResult;
 }
 std::string BoundColumnRefExpression::toDotGraph(uint32_t depth, NodeIdGenerator& idGen) {
