+++ src/compiler/frontend/sql_analyzer.cpp
@@ -167,8 +167,6 @@ std::shared_ptr<ast::TableProducer> SQLCanonicalizer::canonicalize(std::shared_p
                int i = 0;
                //Canonicalize target expressions
                std::ranges::transform(selectNode->targets, selectNode->targets.begin(), [&](std::shared_ptr<ast::ParsedExpression>& target) {
-                  auto it = context->currentScope->groupedByExpressions.find(target);
-
                   return canonicalizeParsedExpression(target, context, true, context->currentScope->extendNodeAfterAggregations);
                });
                //Canonicalize distinct expressions list
@@ -177,12 +175,12 @@ std::shared_ptr<ast::TableProducer> SQLCanonicalizer::canonicalize(std::shared_p
                      return canonicalizeParsedExpression(target, context, true, context->currentScope->extendNodeAfterAggregations);
                   });
                }
-               return pipeOp;
+              break;
             }
             case ast::PipeOperatorType::WHERE: {
                assert(pipeOp->node->nodeType == ast::NodeType::EXPRESSION);
                pipeOp->node = canonicalizeParsedExpression(std::static_pointer_cast<ast::ParsedExpression>(pipeOp->node), context, false);
-               return pipeOp;
+               break;
             }
             case ast::PipeOperatorType::AGGREGATE: {
                auto aggNode = std::dynamic_pointer_cast<ast::AggregationNode>(pipeOp->node);
@@ -208,7 +206,7 @@ std::shared_ptr<ast::TableProducer> SQLCanonicalizer::canonicalize(std::shared_p
                   aggNode->groupByNode->groupByExpressions = std::move(newGroupByExpressions);
                }
 
-               return pipeOp;
+               break;
             }
 
             case ast::PipeOperatorType::RESULT_MODIFIER: {
@@ -223,7 +221,6 @@ std::shared_ptr<ast::TableProducer> SQLCanonicalizer::canonicalize(std::shared_p
 
             default: return pipeOp;
          }
-
          return pipeOp;
       }
       case ast::NodeType::TABLE_REF: {
@@ -668,6 +665,8 @@ std::shared_ptr<ast::TableProducer> SQLQueryAnalyzer::analyzeTableProducer(std::
          switch (queryNode->type) {
             case ast::QueryNodeType::CTE_NODE: {
                auto cteNode = std::static_pointer_cast<ast::CTENode>(queryNode);
+               auto boundCteNode = drv.nf.node<ast::BoundCTENode>(cteNode->loc);
+               boundCteNode->alias = cteNode->alias;
 
                if (cteNode->query) {
                   ast::TargetInfo targetInfo{};
@@ -676,17 +675,16 @@ std::shared_ptr<ast::TableProducer> SQLQueryAnalyzer::analyzeTableProducer(std::
                      auto defineScope = context->createDefineScope();
                      context->pushNewScope();
                      auto subQueryScope = context->currentScope;
-                     cteNode->query = analyzeTableProducer(cteNode->query, context, subQueryResolverScope);
+                     boundCteNode->query = analyzeTableProducer(cteNode->query, context, subQueryResolverScope);
                      targetInfo = context->currentScope->targetInfo;
-                     auto evalBefore = context->currentScope->evalBefore;
                      context->popCurrentScope();
 
-                     cteNode->subQueryScope = *subQueryScope;
+                     boundCteNode->subQueryScope = *subQueryScope;
                      std::vector<std::pair<std::shared_ptr<ast::NamedResult>, std::shared_ptr<ast::NamedResult>>> renamedResults;
                      size_t i = 0;
                      for (auto targetColumns : targetInfo.targetColumns) {
                         auto from = targetColumns;
-                        auto to = std::make_shared<ast::NamedResult>(from->type, context->getUniqueScope(cteNode->alias), from->resultType, from->name);
+                        auto to = std::make_shared<ast::NamedResult>(context->getUniqueScope(cteNode->alias), from->resultType, from->name);
                         to->displayName = from->displayName;
                         if (cteNode->columnNames.size() > i) {
                            to->displayName = cteNode->columnNames[i];
@@ -695,16 +693,16 @@ std::shared_ptr<ast::TableProducer> SQLQueryAnalyzer::analyzeTableProducer(std::
                         renamedResults.emplace_back(std::pair{from, to});
                         i++;
                      }
-                     cteNode->renamedResults = std::move(renamedResults);
+                     boundCteNode->renamedResults = std::move(renamedResults);
 
 
-                     context->ctes.insert({cteNode->alias, {targetInfo, cteNode}});
+                     context->ctes.insert({cteNode->alias, {targetInfo, boundCteNode}});
                   }
                }
                if (cteNode->child) {
-                  cteNode->child = analyzeTableProducer(cteNode->child, context, resolverScope);
+                  boundCteNode->child = analyzeTableProducer(cteNode->child, context, resolverScope);
                }
-               return cteNode;
+               return boundCteNode;
             }
             case ast::QueryNodeType::SET_OPERATION_NODE: {
                auto setOperationNode = std::static_pointer_cast<ast::SetOperationNode>(rootNode);
@@ -743,7 +741,7 @@ std::shared_ptr<ast::TableProducer> SQLQueryAnalyzer::analyzeTableProducer(std::
                   leftColumn->resultType = commonTypes[0];
                   rightColumn->resultType = commonTypes[1];
                   auto commonType = SQLTypeUtils::getCommonType(leftColumn->resultType, rightColumn->resultType);
-                  auto newNamedResult = std::make_shared<ast::NamedResult>(leftColumn->type, newScopeName, commonType, leftColumn->name);
+                  auto newNamedResult = std::make_shared<ast::NamedResult>(newScopeName, commonType, leftColumn->name);
 
                   newNamedResult->displayName = leftColumn->displayName;
                   newTargetInfos.emplace_back(newNamedResult);
@@ -931,6 +929,7 @@ std::shared_ptr<ast::TableProducer> SQLQueryAnalyzer::analyzePipeOperator(std::s
          std::vector<std::shared_ptr<ast::NamedResult>> groupNamedResults{};
          std::vector<std::shared_ptr<ast::BoundExpression>> evalBeforeAggr;
          if (aggregationNode->groupByNode) {
+
             std::ranges::transform(aggregationNode->groupByNode->groupByExpressions, std::back_inserter(groupNamedResults), [&](auto expr) {
                auto boundExpression = analyzeExpression(expr, context, resolverScope);
                assert(boundExpression->namedResult.has_value());
@@ -938,8 +937,8 @@ std::shared_ptr<ast::TableProducer> SQLQueryAnalyzer::analyzePipeOperator(std::s
                switch (boundExpression->exprClass) {
                   case ast::ExpressionClass::BOUND_FUNCTION: {
                      assert(boundExpression->resultType.has_value());
-                     auto scope = boundExpression->alias.empty() ? boundExpression->alias : createTmpScope();
-                     auto n = std::make_shared<ast::NamedResult>(ast::NamedResultType::EXPRESSION, scope, boundExpression->resultType.value(), createTmpScope());
+                     auto scope = boundExpression->alias.empty() ? boundExpression->alias : context->getUniqueScope("tmp_attr");
+                     auto n = std::make_shared<ast::NamedResult>( scope, boundExpression->resultType.value(),  context->getUniqueScope("tmp_attr"));
                      n->displayName = boundExpression->alias.empty() ? "" : boundExpression->alias;
                      evalBeforeAggr.emplace_back(boundExpression);
                      ast::NodeIdGenerator g{};
@@ -950,8 +949,16 @@ std::shared_ptr<ast::TableProducer> SQLQueryAnalyzer::analyzePipeOperator(std::s
                   default:;
                }
 
+
                return boundExpression->namedResult.value();
             });
+            //TODO
+            /*for (auto& n : context->definedAttributes.top()) {
+               n.second->available = false;
+            }
+            for (auto x: groupNamedResults) {
+               x->available = true;
+            }*/
          }
 
          std::vector<std::shared_ptr<ast::BoundFunctionExpression>> boundAggregationExpressions{};
@@ -962,7 +969,7 @@ std::shared_ptr<ast::TableProducer> SQLQueryAnalyzer::analyzePipeOperator(std::s
             //boundExpr->namedResult = std::make_shared<ast::FunctionInfo>("boundExpr->scope", "boundExpr->alias", boundExpr->resultType.value());
             return std::static_pointer_cast<ast::BoundFunctionExpression>(boundExpr);
          });
-         auto mapName = createMapName();
+         auto mapName = context->getUniqueScope("ma");
          auto boundGroupByNode = drv.nf.node<ast::BoundGroupByNode>(aggregationNode->groupByNode ? aggregationNode->groupByNode->loc : aggregationNode->loc, groupNamedResults, aggregationNode->groupByNode ? aggregationNode->groupByNode->groupingSet : std::vector<std::set<size_t>>{});
          std::vector<std::shared_ptr<ast::BoundExpression>> toMap{};
          for (auto& aggr : boundAggregationExpressions) {
@@ -970,8 +977,8 @@ std::shared_ptr<ast::TableProducer> SQLQueryAnalyzer::analyzePipeOperator(std::s
                continue;
             }
             toMap.emplace_back(aggr->arguments[0]);
-            aggr->arguments[0]->alias = createTmpScope();
-            aggr->arguments[0]->namedResult = std::make_shared<ast::NamedResult>(ast::NamedResultType::EXPRESSION, mapName, aggr->arguments[0]->resultType.value(), aggr->arguments[0]->alias);
+            aggr->arguments[0]->alias =  context->getUniqueScope("tmp_attr");
+            aggr->arguments[0]->namedResult = std::make_shared<ast::NamedResult>(mapName, aggr->arguments[0]->resultType.value(), aggr->arguments[0]->alias);
          }
          //ADD to TargetInfo, see Google PIPE sql paper!
          //Maybe Not the best way!
@@ -1011,13 +1018,13 @@ std::shared_ptr<ast::TableProducer> SQLQueryAnalyzer::analyzePipeOperator(std::s
                   if (groupingSet.contains(j)) {
 
                      localGroupBy.emplace_back(groupNamedResults[j]);
-                     auto mappedNamedResult = std::make_shared<ast::NamedResult>(groupNamedResults[j]->type, "rollup_" + std::to_string(rollupId), groupNamedResults[j]->resultType, "tmp_" + std::to_string(j));
+                     auto mappedNamedResult = std::make_shared<ast::NamedResult>("rollup_" + std::to_string(rollupId), groupNamedResults[j]->resultType, "tmp_" + std::to_string(j));
                      mappedNamedResult->displayName = groupNamedResults[j]->displayName;
                      mapToNull.emplace_back(mappedNamedResult);
                      mappedNamedResult->resultType.isNullable = true;
                   } else {
                      present |= (1 << j);
-                     auto mappedNamedResult = std::make_shared<ast::NamedResult>(groupNamedResults[j]->type, "rollup_" + std::to_string(rollupId), groupNamedResults[j]->resultType, "tmp_" + std::to_string(j));
+                     auto mappedNamedResult = std::make_shared<ast::NamedResult>( "rollup_" + std::to_string(rollupId), groupNamedResults[j]->resultType, "tmp_" + std::to_string(j));
                      notAvailable.emplace_back(mappedNamedResult);
                      mappedNamedResult->displayName = groupNamedResults[j]->displayName;
                      mappedNamedResult->resultType.isNullable = true;
@@ -1027,7 +1034,7 @@ std::shared_ptr<ast::TableProducer> SQLQueryAnalyzer::analyzePipeOperator(std::s
 
                std::vector<std::shared_ptr<ast::NamedResult>> aggregationNamedResults{};
                for (auto& aggr : boundAggrNode->aggregations) {
-                  auto namedResultAggr = std::make_shared<ast::NamedResult>(aggr->namedResult.value()->type, "rollupAgg_" + std::to_string(rollupId), aggr->namedResult.value()->resultType, aggr->namedResult.value()->name);
+                  auto namedResultAggr = std::make_shared<ast::NamedResult>("rollupAgg_" + std::to_string(rollupId), aggr->namedResult.value()->resultType, aggr->namedResult.value()->name);
                   namedResultAggr->displayName = aggr->namedResult.value()->displayName;
                   aggregationNamedResults.emplace_back(namedResultAggr);
                }
@@ -1038,7 +1045,7 @@ std::shared_ptr<ast::TableProducer> SQLQueryAnalyzer::analyzePipeOperator(std::s
                boundAggrNode->groupByNode->localGroupByNamedResults.emplace_back(std::move(localGroupBy));
                boundAggrNode->groupByNode->localMapToNullNamedResults.emplace_back(std::move(mapToNull));
                boundAggrNode->groupByNode->localNotAvailableNamedResults.emplace_back(std::move(notAvailable));
-               auto presentNamedResult = std::make_shared<ast::NamedResult>(ast::NamedResultType::EXPRESSION, boundAggrNode->mapName, catalog::Type::int64(), "intval" + std::to_string(present));
+               auto presentNamedResult = std::make_shared<ast::NamedResult>(boundAggrNode->mapName, catalog::Type::int64(), "intval" + std::to_string(present));
                boundAggrNode->groupByNode->localPresentIntval.emplace_back(std::pair{present,presentNamedResult});
                rollupId++;
 
@@ -1077,7 +1084,7 @@ std::shared_ptr<ast::TableProducer> SQLQueryAnalyzer::analyzePipeOperator(std::s
                for (size_t j = 0; j < currentLocalAttributes.size(); j++) {
                   auto left = currentAttributes[j];
                   auto right = currentLocalAttributes[j];
-                  auto unionNamedResult = std::make_shared<ast::NamedResult>(left->type, rollUpUnionName + std::to_string(i), right->resultType, left->name);
+                  auto unionNamedResult = std::make_shared<ast::NamedResult>(rollUpUnionName + std::to_string(i), right->resultType, left->name);
                   unionNamedResults.emplace_back(unionNamedResult);
                   unionNamedResult->displayName = left->displayName;
 
@@ -1152,7 +1159,7 @@ std::shared_ptr<ast::TableProducer> SQLQueryAnalyzer::analyzePipeOperator(std::s
 
             return boundExpression;
          });
-         auto mapName = createMapName();
+         auto mapName = context->getUniqueScope("map");
 
          for (auto& parsedExpression: boundExtensions) {
             //TODO very similar to the logic in Select. Maybe combine both
@@ -1189,7 +1196,7 @@ std::shared_ptr<ast::TableProducer> SQLQueryAnalyzer::analyzePipeOperator(std::s
                   if (resultType.useZeroInsteadOfNull) {
                      resultType.isNullable = false;
                   }
-                  auto n = std::make_shared<ast::NamedResult>(ast::NamedResultType::EXPRESSION, scope, resultType , createTmpScope());
+                  auto n = std::make_shared<ast::NamedResult>(scope, resultType ,  context->getUniqueScope("tmp_attr"));
                   n->displayName = parsedExpression->alias.empty() ? "" : parsedExpression->alias;
                   context->mapAttribute(resolverScope, parsedExpression->alias.empty() ? n->name : parsedExpression->alias, n);
                   if (extendNode->hidden) {
@@ -1200,8 +1207,6 @@ std::shared_ptr<ast::TableProducer> SQLQueryAnalyzer::analyzePipeOperator(std::s
 
 
 
-
-                  //context->currentScope->evalBefore.emplace_back(parsedExpression);
 
                   parsedExpression->namedResult = n;
 
@@ -1244,7 +1249,7 @@ std::shared_ptr<ast::TableProducer> SQLQueryAnalyzer::analyzeTableRef(std::share
                std::vector<std::shared_ptr<ast::NamedResult>> namedResults{};
 
                std::ranges::transform(cteNode->renamedResults, std::back_inserter(namedResults), [&](auto& pair) {
-                  auto namedResult = std::make_shared<ast::NamedResult>(pair.second->type, context->getUniqueScope(baseTableRef->tableName), pair.second->resultType, pair.second->name);
+                  auto namedResult = std::make_shared<ast::NamedResult>(context->getUniqueScope(baseTableRef->tableName), pair.second->resultType, pair.second->name);
                   namedResult->displayName = pair.second->displayName;
                   return namedResult;
                });
@@ -1364,7 +1369,6 @@ std::shared_ptr<ast::TableProducer> SQLQueryAnalyzer::analyzeTableRef(std::share
                   boundCondition = analyzeExpression(std::get<std::shared_ptr<ast::ParsedExpression>>(join->condition), context, resolverScope);
 
                }
-               //TODO
 
                std::vector<std::pair<std::shared_ptr<ast::NamedResult>, std::shared_ptr<ast::NamedResult>>> outerJoinMapping;
                std::string outerjoinName;
@@ -1377,7 +1381,7 @@ std::shared_ptr<ast::TableProducer> SQLQueryAnalyzer::analyzeTableRef(std::share
                      if (it == remapped.end()) {
                         auto scope = x.second->scope;
                         auto name = x.second->name;
-                        auto namedResult = std::make_shared<ast::NamedResult>(x.second->type, outerjoinName, x.second->resultType, name);
+                        auto namedResult = std::make_shared<ast::NamedResult>(outerjoinName, x.second->resultType, name);
 
                         //Make mapping output nullable
                         namedResult->resultType.isNullable = true;
@@ -1426,10 +1430,6 @@ std::shared_ptr<ast::TableProducer> SQLQueryAnalyzer::analyzeTableRef(std::share
                   mapping.insert(mapping.end(), localMapping.begin(), localMapping.end());
                }
 
-
-
-
-
                std::shared_ptr<ast::BoundExpression> boundCondition;
                {
                   auto predScope = context->createResolverScope();
@@ -1443,8 +1443,6 @@ std::shared_ptr<ast::TableProducer> SQLQueryAnalyzer::analyzeTableRef(std::share
                   }
                   boundCondition = analyzeExpression(std::get<std::shared_ptr<ast::ParsedExpression>>(join->condition), context, resolverScope);
                }
-               //TODO
-
                std::vector<std::pair<std::shared_ptr<ast::NamedResult>, std::shared_ptr<ast::NamedResult>>> outerJoinMapping;
                std::string outerjoinName;
                static size_t id = 0;
@@ -1457,7 +1455,7 @@ std::shared_ptr<ast::TableProducer> SQLQueryAnalyzer::analyzeTableRef(std::share
                      if (it == remapped.end()) {
                         auto scope = x.second->scope;
                         auto name = x.second->name + "_" +  std::to_string(id++);
-                        auto namedResult = std::make_shared<ast::NamedResult>(x.second->type,  outerjoinName, x.second->resultType,   name);
+                        auto namedResult = std::make_shared<ast::NamedResult>(outerjoinName, x.second->resultType,   name);
 
                         //Make mapping output nullable
                         namedResult->resultType.isNullable = true;
@@ -1497,7 +1495,6 @@ std::shared_ptr<ast::TableProducer> SQLQueryAnalyzer::analyzeTableRef(std::share
             subQueryScope = context->currentScope;
             t = analyzeTableProducer(subquery->subSelectNode, context, subQueryResolverScope);
             targetInfo = context->currentScope->targetInfo;
-            evalBefore = context->currentScope->evalBefore;
             context->popCurrentScope();
          }
          size_t i = 0;
@@ -1554,10 +1551,10 @@ std::shared_ptr<ast::TableProducer> SQLQueryAnalyzer::analyzeTableRef(std::share
             return t;
          });
          std::vector<std::shared_ptr<ast::NamedResult>> namedResults{};
-         auto scope = createTmpScope();
+         auto scope = context->getUniqueScope("tmp_attr");
          for (size_t i = 0; i < commonTypes.size(); i++) {
             auto name = "const_" + std::to_string(i);
-            auto namedResult = std::make_shared<ast::NamedResult>(ast::NamedResultType::EXPRESSION, scope, commonTypes[i], name);
+            auto namedResult = std::make_shared<ast::NamedResult>(scope, commonTypes[i], name);
             namedResults.push_back(namedResult);
             context->currentScope->targetInfo.add(namedResult);
          }
@@ -1834,7 +1831,7 @@ std::shared_ptr<ast::BoundExpression> SQLQueryAnalyzer::analyzeExpression(std::s
          auto function = std::static_pointer_cast<ast::FunctionExpression>(rootNode);
          std::vector<std::shared_ptr<ast::BoundExpression>> boundArguments{};
          if (rootNode->type == ast::ExpressionType::AGGREGATE) {
-            auto scope = createTmpScope();
+            auto scope =  context->getUniqueScope("tmp_attr");
             auto fName = function->alias.empty() ? function->functionName : function->alias;
             std::ranges::transform(function->functionName, function->functionName.begin(), ::toupper);
             std::shared_ptr<ast::BoundFunctionExpression> boundFunctionExpression = nullptr;
@@ -1940,7 +1937,7 @@ std::shared_ptr<ast::BoundExpression> SQLQueryAnalyzer::analyzeExpression(std::s
                error("Not implemented", rootNode->loc);
             }
 
-            auto fInfo = std::make_shared<ast::NamedResult>(ast::NamedResultType::Function, scope, resultType, fName);
+            auto fInfo = std::make_shared<ast::NamedResult>(scope, resultType, fName);
 
             fInfo->displayName = function->alias;
             context->mapAttribute(resolverScope, fName, fInfo);
@@ -1949,7 +1946,7 @@ std::shared_ptr<ast::BoundExpression> SQLQueryAnalyzer::analyzeExpression(std::s
             return boundFunctionExpression;
          }
          std::ranges::transform(function->functionName, function->functionName.begin(), ::toupper);
-         auto scope = createTmpScope();
+         auto scope =  context->getUniqueScope("tmp_attr");
          auto fName = function->alias.empty() ? function->functionName : function->alias;
          std::shared_ptr<ast::BoundFunctionExpression> boundFunctionExpression = nullptr;
          catalog::NullableType resultType{catalog::Type::noneType()};
@@ -2110,7 +2107,7 @@ std::shared_ptr<ast::BoundExpression> SQLQueryAnalyzer::analyzeExpression(std::s
          if (boundFunctionExpression == nullptr) {
             error("Function '" << function->functionName << "' not implemented", function->loc);
          }
-         auto fInfo = std::make_shared<ast::NamedResult>(ast::NamedResultType::Function, scope, resultType, fName);
+         auto fInfo = std::make_shared<ast::NamedResult>(scope, resultType, fName);
 
 
          boundFunctionExpression->namedResult = fInfo;
@@ -2233,7 +2230,7 @@ std::shared_ptr<ast::BoundExpression> SQLQueryAnalyzer::analyzeExpression(std::s
          }
 
          auto namedResult = subqueryTargetInfo.targetColumns[0];
-         auto x = std::make_shared<ast::NamedResult>(namedResult->type, namedResult->scope, namedResult->resultType, namedResult->name);
+         auto x = std::make_shared<ast::NamedResult>(namedResult->scope, namedResult->resultType, namedResult->name);
          auto resultType = namedResult->resultType;
          //TODO check for correctness: Is every subquery nullable
          resultType.isNullable = true;
@@ -2434,7 +2431,7 @@ std::shared_ptr<ast::BoundColumnRefExpression> SQLQueryAnalyzer::analyzeColumnRe
       error("Column not found", columnRef->loc);
    }
    found->displayName = columnRef->alias.empty() ? found->displayName : columnRef->alias;
-   return drv.nf.node<ast::BoundColumnRefExpression>(columnRef->loc, found->scope, found->resultType, found, columnRef->alias);
+   return drv.nf.node<ast::BoundColumnRefExpression>(columnRef->loc, found->resultType, found, columnRef->alias);
 }
 
 /*
